+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## upcoming release
 ENHANCEMENTS:
 * add resource junos_security_utm_custom_url_pattern
+* add resource junos_security_utm_policy
 * add resource junos_security_utm_profile_web_filtering_juniper_enhanced
 
 BUG FIXES: